requirements.txt
openai>=1.10.0
langchain==0.1.11
langchain-openai==0.1.1
python-dotenv==1.0.0
pydantic<2.0.0,>=1.10.8
flask==2.3.3
flask-socketio==5.3.6 